feat: add store.dart
